+++ compiler/syntax/src/res_parens.ml
@@ -140,12 +140,12 @@ let binary_expr_operand ~is_lhs expr =
       else Nothing)
 
 let sub_binary_expr_operand parent_operator child_operator =
-  let prec_parent = ParsetreeViewer.operator_precedence parent_operator in
-  let prec_child = ParsetreeViewer.operator_precedence child_operator in
+  let open ParsetreeViewer in
+  let prec_parent = operator_precedence parent_operator in
+  let prec_child = operator_precedence child_operator in
   prec_parent > prec_child
-  || prec_parent == prec_child
-     && not
-          (ParsetreeViewer.flattenable_operators parent_operator child_operator)
+  || is_equality_operator parent_operator
+     && is_equality_operator child_operator
   || (* a && b || c, add parens to (a && b) for readability, who knows the difference by heart… *)
   (parent_operator = "||" && child_operator = "&&")