+++ tests/tests/src/unified_ops_test.mjs
@@ -59,6 +59,12 @@ function odd(n) {
   return n % 2 === 1;
 }
 
+let pow1 = 4;
+
+let pow2 = 2 ** 2;
+
+let pow3 = 2n ** 2n;
+
 let int = 3;
 
 export {
@@ -79,5 +85,8 @@ export {
   case2,
   even,
   odd,
+  pow1,
+  pow2,
+  pow3,
 }
 /* No side effect */