+++ tests/tests/src/variantsMatching.mjs
@@ -337,9 +337,9 @@ let MyNullableExtended = {
 function area(shape) {
   switch (shape.kind) {
     case 1 :
-      return Math.PI * Math.pow(shape.radius, 2);
+      return Math.PI * shape.radius ** 2;
     case "square" :
-      return Math.pow(shape.sideLength, 2);
+      return shape.sideLength ** 2;
     case "rectangle" :
       return shape.width * shape.height;
   }